ModalDialog includes now a modal login window
Add a new rendered DOM for an iframe with the login page. And two functions: one (setwin)  for get the window value in the iframe element (and write it in the winOb state) and other (getwin) that watches the changes on the uri of this window and close if the process has ended, that is, when windows reach the callback page from the form.